+++ app/src/js/components/Executions/execution-snapshot.js
@@ -2,6 +2,7 @@ import React, { useEffect, useMemo } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { Alert } from 'react-bootstrap';
+import isEmpty from 'lodash/isEmpty';
 import SortableTable from '../SortableTable/SortableTable';
 import { getExecutionStatus } from '../../actions';
 import { formatEvents, subColumns } from '../../utils/table-config/executions';
@@ -19,7 +20,7 @@ const ExecutionSnapshot = ({
   const hasEvents = events?.length > 1;
 
   useEffect(() => {
-    if (isExpanded && !currentExecutionStatus) {
+    if (isExpanded && isEmpty(currentExecutionStatus)) {
       dispatch(getExecutionStatus(arn));
     }
   }, [dispatch, arn, isExpanded, currentExecutionStatus]);