+++ CHANGELOG.md
@@ -31,6 +31,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - **CUMULUS-375**
   - Fixed display issues on individual provider page
+- **CUMULUS-2623**
+  - Fixed issue where dropdown styles were not being applied in production
 - **CUMULUS-2721**
   - Fixed issue where boolean environment variables were being read as strings
 - **CUMULUS-2772**